data-analyze-refined_skill/requirements.txt
pandas>=1.5.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0
openpyxl>=3.1.0
python-docx>=0.8.11
Pillow>=10.0.0
plotly>=5.15.0